Commit usuario controlador

+++ src/main/java/org/springframework/samples/petclinic/usuario/UsuarioRepositorio.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioRepositorio extends CrudRepository<Usuario,Long> {
+public interface UsuarioRepositorio extends CrudRepository<Usuario,Integer> {
 	
 	public Collection<Usuario> findAll();
